moved sound effects out of othehr scripts into its own SoundEffects.cs
