+++ Assets/Animations/characterController.controller
@@ -271,9 +271,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d4c7d954a446b5b47bc2b63c6c9ec428, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  movementAudioPlayer: {fileID: 0}
-  Jump_Up_Sound: {fileID: 0}
-  Jump_Down_Sound: {fileID: 0}
   _inactiveTimerMax: 5
   acceleration: 0.1
   deceleration: 0.5
@@ -533,6 +530,9 @@ AnimatorTransition:
   m_Conditions:
   - m_ConditionMode: 1
     m_ConditionEvent: isWalking
+    m_EventTreshold: 0
+  - m_ConditionMode: 3
+    m_ConditionEvent: speed
     m_EventTreshold: 0
   m_DstStateMachine: {fileID: 0}
   m_DstState: {fileID: -4856261896828157139}
@@ -1649,20 +1649,20 @@ AnimatorStateMachine:
     m_Position: {x: -870, y: 280, z: 0}
   - serializedVersion: 1
     m_State: {fileID: -3972526797841380905}
-    m_Position: {x: 270, y: 470, z: 0}
+    m_Position: {x: -60, y: 300, z: 0}
   - serializedVersion: 1
     m_State: {fileID: 3835883710201457506}
-    m_Position: {x: 640, y: 160, z: 0}
+    m_Position: {x: 490, y: 300, z: 0}
   - serializedVersion: 1
     m_State: {fileID: -4856261896828157139}
-    m_Position: {x: -160, y: 170, z: 0}
+    m_Position: {x: -210, y: -220, z: 0}
   m_ChildStateMachines:
   - serializedVersion: 1
     m_StateMachine: {fileID: -1478735247052279545}
-    m_Position: {x: 540, y: -330, z: 0}
+    m_Position: {x: 630, y: -240, z: 0}
   - serializedVersion: 1
     m_StateMachine: {fileID: 4548517076853111044}
-    m_Position: {x: 0, y: -350, z: 0}
+    m_Position: {x: 220, y: -520, z: 0}
   m_AnyStateTransitions: []
   m_EntryTransitions: []
   m_StateMachineTransitions:
