+++ Assets/Prefabs/Female_Player.prefab
@@ -2248,6 +2248,7 @@ GameObject:
   - component: {fileID: 9079016264875487769}
   - component: {fileID: 9079016264875487771}
   - component: {fileID: 7279413962254325220}
+  - component: {fileID: 2824048881423560565}
   m_Layer: 0
   m_Name: Female_Player
   m_TagString: Untagged
@@ -2333,18 +2334,11 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 09cd0a4cca98e0f49a975f972aefd513, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  fightingAudioPlayer: {fileID: 7279413962254325220}
   is_blocking: 0
   is_punching: 0
   is_kicking: 0
   punch_time: 1.733
   kick_time: 1.3
-  Blocking_Sound: {fileID: 8300000, guid: 3fd639473fc13be42b07df2755be05ab, type: 3}
-  Punch_Impact_Sound: {fileID: 8300000, guid: 4f8fbafe2b8651942bdbc006e659b696, type: 3}
-  Punch_Whoosh_Sound: {fileID: 8300000, guid: f0d9d714e04c7604388cd8e938c6f504, type: 3}
-  Kick_Whoosh_Sound: {fileID: 8300000, guid: 6bbbb367e579eed45bad8799fc603898, type: 3}
-  Kick_Impact_Sound: {fileID: 8300000, guid: 2f68415471fa4354eb9d9704f45ac951, type: 3}
-  Jump_Whoosh_Sound: {fileID: 8300000, guid: f958282e5fa2e8d4599de2431096cef9, type: 3}
 --- !u!114 &9079016264875487748
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -2357,9 +2351,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d4c7d954a446b5b47bc2b63c6c9ec428, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  movementAudioPlayer: {fileID: 7279413962254325220}
-  Jump_Up_Sound: {fileID: 8300000, guid: ba7e1a28f3073484a83be000fb673418, type: 3}
-  Jump_Down_Sound: {fileID: 8300000, guid: aac7c2e920f305a4b9125e6399596c82, type: 3}
   _inactiveTimerMax: 5
   acceleration: 0.8
   deceleration: 1.2
@@ -2535,6 +2526,27 @@ AudioSource:
     m_PreInfinity: 2
     m_PostInfinity: 2
     m_RotationOrder: 4
+--- !u!114 &2824048881423560565
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8160044105240706494}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c6f67253c2aa9424d9f517e2eb1b5cbf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  characterAudioPlayer: {fileID: 7279413962254325220}
+  Blocking_Sound: {fileID: 8300000, guid: 3fd639473fc13be42b07df2755be05ab, type: 3}
+  Punch_Impact_Sound: {fileID: 8300000, guid: 4f8fbafe2b8651942bdbc006e659b696, type: 3}
+  Punch_Whoosh_Sound: {fileID: 8300000, guid: f0d9d714e04c7604388cd8e938c6f504, type: 3}
+  Kick_Whoosh_Sound: {fileID: 8300000, guid: 6bbbb367e579eed45bad8799fc603898, type: 3}
+  Kick_Impact_Sound: {fileID: 8300000, guid: 2f68415471fa4354eb9d9704f45ac951, type: 3}
+  Jump_Up_Sound: {fileID: 8300000, guid: ba7e1a28f3073484a83be000fb673418, type: 3}
+  Jump_Down_Sound: {fileID: 8300000, guid: aac7c2e920f305a4b9125e6399596c82, type: 3}
+  Jump_Whoosh_Sound: {fileID: 8300000, guid: f958282e5fa2e8d4599de2431096cef9, type: 3}
 --- !u!1 &8284975885569072869
 GameObject:
   m_ObjectHideFlags: 0
