+++ Assets/Scripts/SoundEffects.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c6f67253c2aa9424d9f517e2eb1b5cbf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
